% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_BC_date.R
\name{as_BC_date}
\alias{as_BC_date}
\title{function to create date BC}
\usage{
as_BC_date(year, month = 1, day = 1)
}
\arguments{
\item{year}{the year BC (either + or -)}

\item{month}{the calender month, if omitted 01 is asumed}

\item{day}{the day of that month, if omitted 01 is assumed}
}
\value{
date as BC date
}
\description{
the lubridate as_date function doesn't handle dates BC correctly,
hence this function creates the BC date by first projecting the date as AD,
and then subtract from 0 AD.
}
\note{
this function is for internal use

the year can be passed as negative or positive, the function assumes
the user needs a BC date
}
\examples{
\dontrun{
   BCDate <- as_BC_date(-2150, 1, 1)
}

}
