% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Function to create popup_label text for interactive map}
\usage{
eq_create_label(data = NULL, line_1 = "LOCATION", line_2 = "EQ_PRIMARY",
  line_3 = "TOTAL_DEATHS")
}
\arguments{
\item{data}{a data.frame containing columns named:
"LOCATION" LOCATION info of earthquake events
"EQ_PRIMARY" size of earthquake events on Richter scale
"TOTAL_DEATHS" Total deaths info of earthquake events}

\item{line_1}{the name of a column in \code{data}; by default "LOCATION"}

\item{line_2}{the name of a column in \code{data}; by default "EQ_PRIMARY"}

\item{line_3}{the name of a column in \code{data}; by default "TOTAL_DEATHS"}
}
\value{
character vector containing popup-text
}
\description{
This function creates labels for an interactive map using three
columns from a data.frame (the \code{data} variable). By default these are the
"LOCATION" (the \code{line_1} variable), "EQ_PRIMARY" (the \code{line_2}
variable) AND "TOTAL_DEATHS" (the \code{line_3} variable) columns. It will
omit all \code{NA} info from the label text, and returns a vector that can be
used as popuptext column in the data.frame as label text in the eq_map function.
}
\examples{
\dontrun{
# assuming earthquake data is in the file earthquakes.tvs.gz in the wd
readr::read_delim("earthquakes.tsv.gz", delim = "\\t") \%>\%
    eq_clean_data() \%>\%
    dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
    dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
    eq_map(annot_col = "popup_text")#'
}

}
