% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Function to clean the location info in noaa earthquake data set}
\usage{
eq_location_clean(raw_location)
}
\arguments{
\item{row_location}{a data.frame containing the LOCATION_NAME column from the
naoo dataset}
}
\value{
data.frame containing cleaned location data in one column
}
\description{
This function creates labels for an interactive map using 
"LOCATION", EQ_PRIMARY" AND "TOTAL_DEATHS" columns from a data.frame 
(\code{data} variable). It will omit all \code{NA} info from the label text,
and returns a single column data.frame with the label text.
}
\note{
this function is for internal use
}
\examples{
\dontrun{
locNameClean <- eq_location_clean(naoo_data$LOCATION_NAME)
}

}
