% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomTimeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Create a timeline diagram}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
Grob representing timeline(s) and datapoints
}
\description{
This geom plots a timeline of dates (requires \code{x} date
aesthetics) in one or more lines (depending on optional \code{y} aesthetics
to group data by selected variable). Furthermore size and colour can be
linked to numeric variables.
}
\note{
required_aes x date of data points to be plotted

default_aes y, size, colour with y=0.1, size=3, colour="grey"
}
\examples{
\dontrun{
eq_data \%>\% eq_clean_data() \%>\%
   filter(COUNTRY \%in\% c("GREECE", "ITALY"), year(DATE) > 2000) \%>\%
   ggplot(aes(x = DATE,
                  y = COUNTRY,
                  color = TOTAL_DEATHS,
                  size = EQ_PRIMARY)) +
   geom_timeline() +
   geom_timeline_label(aes(label = LOCATION), n_max = 8) +
   theme_timeline() +
   scale_size(name = "Richter scale", limits = c(5, 8)) +
   scale_color_gradient(name = "# Deaths", low = "black", high = "blue")
}

}
