% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomTimeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{ggproto class for geom_timeline}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeline
}
\description{
for internal use
}
\keyword{datasets}
