% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAAearthquakeAnalysis.R
\docType{package}
\name{NOAAearthquakeAnalysis}
\alias{NOAAearthquakeAnalysis}
\alias{NOAAearthquakeAnalysis-package}
\title{\code{NOAAearthquakeAnalysis} package}
\description{
See the README on
\href{https://github.com/RMHoek/NOAAearthquakeAnalysis#readme}{GitHub}
}
