% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_YMD.R
\name{unite_YMD}
\alias{unite_YMD}
\title{create single column data.frame with dates based on YMD data}
\usage{
unite_YMD(YMD_frame)
}
\arguments{
\item{YMD_frame}{a data.frame with three columns:
"YEAR" the year of the event
"MONTH" the month of the year of the event or NA
"DAY" the day of the month of the event or NA}
}
\value{
data.frame with single column containing DATE info in date format
}
\description{
this function creates a data.frame with a single column containing
dates based on a data.frame containing three columns "YEAR", "MONTH", "DAY".
}
\note{
this function is for internal use

if Month or Day are NA they will be set to 1
}
\examples{
\dontrun{
    datesFrame <- naoo_data \%>\% dplyr::select(YEAR, MONTH, DAY)
    datesFrame <- unite_YMD(datesFrame)
}

}
