% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Function to clean the location info in noaa earthquake data set}
\usage{
eq_location_clean(raw_location)
}
\arguments{
\item{raw_location}{a single column data.frame containing the LOCATION_NAME
column from the naoo dataset}
}
\value{
data.frame containing cleaned location data in one column
}
\description{
This function cleans the info in the original NOAA data.frame
column "LOCATION_NAME", stripping redundant info like country name and
secundary location descriptions. It returns a single column data.frame with
the location text.
}
\note{
this function is for internal use
}
\examples{
\dontrun{
locNameClean <- eq_location_clean(naoo_data$LOCATION_NAME)
}

}
